+++ data/sales/raw_data/gen.txt
@@ -1,0 +1,160 @@
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L1','L2']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L3','L4']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L5','L6']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L7','L8']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L209','L210']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L9','L10']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L11','L12']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L13','L14']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L15','L16']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L17','L18']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L19','L20']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L21','L22','L23','L24','L25','L26','L27','L28']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L29','L30']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L261','L262']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L263','L264']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L31','L32']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L33','L34']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L35','L36']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L317','L318']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L319','L320']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L321','L322']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L323','L324']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L325','L326']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L327','L328']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L329','L330']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L37','L38']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L39','L40']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L41','L42']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L43','L44']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L45','L46']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L47','L48']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L49','L50']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L51','L52']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L53','L54']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L55','L56']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L57','L58']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L59','L60']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L61','L62']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L63','L64']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L65','L66']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L67','L68']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L69','L70']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L71','L72']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L73','L74']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L75','L76']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L77','L78']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L79','L80']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L81','L82']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L83','L84']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L85','L86']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L87','L88']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L89','L90']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L91','L92']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L209','L210']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L211','L212']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L213','L214']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L215','L216']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L217','L218']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L219','L220']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L221','L222']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L223','L224']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L225','L226']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L227','L228']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L229','L230']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L231','L232']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L233','L234']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L235','L236']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L237','L238']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L239','L240']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L241','L242']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L243','L244']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L245','L246']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L247','L248']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L249','L250']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L251','L252']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L253','L254']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L255','L256']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L257','L258']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L259','L260']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L93','L94']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L95','L96']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L97','L98']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L99','L100']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L101','L102']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L103','L104']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L105','L106']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L107','L108']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L109','L110']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L111','L112']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L113','L114']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L115','L116']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L117','L118']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L119','L120']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L121','L122']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L123','L124']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L125','L126']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L127','L128']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L129','L130']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L131','L132']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L133','L134']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L135','L136']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L137','L138']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L139','L140']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L265','L266']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L267','L268','L269','L270']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L271','L272']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L273','L274']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L275','L276']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L277','L278']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L279','L280']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L281','L282']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L283','L284']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L285','L286']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L287','L288']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L289','L290']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L291','L292']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L293','L294']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L295','L296']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L297','L298']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L299','L300']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L301','L302']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L303','L304']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L305','L306']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L307','L308']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L309','L310']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L311','L312']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L313','L314']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L315','L316']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L141','L142','L143','L144']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L145','L146']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L147','L148']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L149','L150']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L151','L152']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L153','L154']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L155','L156']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L157','L158']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L159','L160']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L161','L162']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L163','L164']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L165','L166']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L167','L168']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L169','L170']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L171','L172']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L173','L174']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L175','L176']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L177','L178']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L179','L180']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L181','L182']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L183','L184']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L185','L186']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L187','L188']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L189','L190']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L191','L192']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L193','L194']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L195','L196']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L197','L198']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L199','L200']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L201','L202']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L203','L204']
+u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L205','L206']
