+++ requirements.txt
@@ -1,0 +1,7 @@
+scikit-learn
+tensorflow
+tensorlayer
+numpy
+click
+tqdm
+nltk
