--- seq2seq-chatbot/.gitignore
@@ -1,18 +1,0 @@
-dea
-docs/_build
-tensorlayer
-tensorlayer/__pacache__
-tensorlayer/.DS_Store
-.DS_Store
-dist
-build/
-tensorlayer.egg-info
-data/.DS_Store
-*.pyc
-*.gz
-.spyproject/
-.vscode/*
-
-env/
-venv/
-.idea/
