--- seq2seq-chatbot/data/__init__.py
@@ -1,6 +1,0 @@
-
-from __future__ import absolute_import
-
-# from . import twitter
-# from . import imagenet_classes
-# from . import
