--- seq2seq-chatbot/data/cornell_corpus/0.txt
@@ -1,4 +1,0 @@
-Can we make this quick?  Roxanne Korrine and Andrew Barrett are having an incredibly horrendous public break- up on the quad.  Again.
-Well, I thought we'd start with pronunciation, if that's okay with you.
-Not the hacking and gagging and spitting part.  Please.
-Okay... then how 'bout we try out some French cuisine.  Saturday?  Night?
