--- seq2seq-chatbot/data/cornell_corpus/1.txt
@@ -1,2 +1,0 @@
-You're asking me out.  That's so cute. What's your name again?
-Forget it.
