--- seq2seq-chatbot/data/cornell_corpus/2.txt
@@ -1,4 +1,0 @@
-No, no, it's my fault -- we didn't have a proper introduction ---
-Cameron.
-The thing is, Cameron -- I'm at the mercy of a particularly hideous breed of loser.  My sister.  I can't date until she does.
-Seems like she could get a date easy enough...
