--- seq2seq-chatbot/data/cornell_corpus/3.txt
@@ -1,3 +1,0 @@
-Why?
-Unsolved mystery.  She used to be really popular when she started high school, then it was just like she got sick of it or something.
-That's a shame.
