--- seq2seq-chatbot/data/cornell_corpus/4.txt
@@ -1,2 +1,0 @@
-Gosh, if only we could find Kat a boyfriend...
-Let me see what I can do.
