--- seq2seq-chatbot/data/cornell_corpus/5.txt
@@ -1,5 +1,0 @@
-C'esc ma tete. This is my head
-Right.  See?  You're ready for the quiz.
-I don't want to know how to say that though.  I want to know useful things. Like where the good stores are.  How much does champagne cost?  Stuff like Chat.  I have never in my life had to point out my head to someone.
-That's because it's such a nice one.
-Forget French.
