--- seq2seq-chatbot/data/cornell_corpus/6.txt
@@ -1,2 +1,0 @@
-How is our little Find the Wench A Date plan progressing?
-Well, there's someone I think might be --
