--- seq2seq-chatbot/data/cornell_corpus/7.txt
@@ -1,2 +1,0 @@
-There.
-Where?
