--- seq2seq-chatbot/data/cornell_corpus/8.txt
@@ -1,2 +1,0 @@
-You got something on your mind?
-I counted on you to help my cause. You and that thug are obviously failing. Aren't we ever going on our date?
