--- seq2seq-chatbot/data/cornell_corpus/9.txt
@@ -1,2 +1,0 @@
-You have my word.  As a gentleman
-You're sweet.
