--- seq2seq-chatbot/data/cornell_corpus/raw_data/movie_conversations.txt
@@ -1,29 +1,0 @@
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L1', 'L2']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L3', 'L4']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L5', 'L6']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L7', 'L8']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L9','L10']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L11','L12']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L13','L14']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L15','L16']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L17','L18']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L19','L20']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L21', 'L22','L23','L24','L25','L26','L27','L28','L29','L30','L31','L32']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L33', 'L34','L35','L36','L37','L38','L39','L40']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L41', 'L42','L43','L44']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L45', 'L46','L47','L48','L49','L50']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L51', 'L52']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L53', 'L56']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L57', 'L58','L59','L60','L61']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L62', 'L63','L64','L65','L66','L67','L68','L69']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L70', 'L71']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L72', 'L73']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L74', 'L75']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L76', 'L77']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L78', 'L79']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L80', 'L81']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L82', 'L83']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L84', 'L85']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L86', 'L87']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L88', 'L89']
-u0 +++$+++ u2 +++$+++ m0 +++$+++ ['L90', 'L91','L92','L93']
