--- seq2seq-chatbot/data/sales/data.py
@@ -1,23 +1,0 @@
-#! /usr/bin/python
-# -*- coding: utf-8 -*-
-
-EN_WHITELIST = '0123456789abcdefghijklmnopqrstuvwxyzôâăơưê ' # space is included in whitelist
-EN_BLACKLIST = '!"#$%&\'()*+,-./:;<=>?@[\\]^_`{|}~\''
-
-limit = {
-        'maxq' : 25,
-        'minq' : 2,
-        'maxa' : 25,
-        'mina' : 2
-        }
-
-UNK = 'unk'
-import random
-
-import nltk         # lib for statistics nlp (tokenization, tag....)
-import itertools    # cycle iteration
-from collections import defaultdict
-
-import numpy as np
-
-import pickle
