--- seq2seq-chatbot/data/sales/sales.txt
@@ -1,37 +1,0 @@
-Salesman: Sir, how may I help you?
-Customer: I’m looking for a pair of leather shoes.
-Salesman: Hello sir, welcome to our store.
-Customer: Thank you.
-Salesman: How can I help you, sir.
-Customer: Actually, I am planning to buy an air conditioner, so could you
-please help me with that.
-Salesman: Sure, sir. Could you please specify me the brand so that I can be more transparent with the details?
-Customer: I am using Panasonic AC these days but I am not that satisfied with its services.
-Salesman: You should try the LG air conditioner. It is already in high demand.
-Customer: What kind of service does it offer?
-Salesman: In LG air conditioner, you will get a two-year warranty. It also has five free services. The installation is free. The maintenance is also taken care by the company.
-Customer: That’s perfect.
-Salesman: Sir, where do you want to install the air conditioner?
-Customer: In the living hall.
-Salesman: How much Tonne is required?
-Customer: 2.0 Tonne would be better.
-Salesman: Sure sir.
-Customer: Could you please tell me the price of this air conditioner and the mode of payment?
-Salesman: It will cost you around 30000/- INR including service tax. And you can pay by card, cash or cheque.
-Customer: Perfect, I would like to pay by card.
-Salesman: Sure, sir. Let’s complete the formalities and proceed further.
-Customer: Okay!
-Salesman:  Good Morning. May I Help You, Please?
-Customer: Yes I Want To Buy Some Books.
-Salesman: There Are Different Kinds Of Books Here. What Type Of Books Do You Want?
-Customer: I Want Story Books For The Children.
-Salesman: Yes, I Want To Buy Children’s Book.
-Salesman: You’ll Find Children’s Book in Those Selves. Anyhow, I Am Bringing Them To You. They Are Very Popular Among The Children.
-Customer: Please, Bring.
-Salesman: Here Are The Books.
-Customer: How Much Do They Cost?
-Salesman: Two Hundred.
-Customer: Here Is The Money.
-Salesman: Thank You. Please, Visit Us Next Time.
-Customer: You’re Most Welcome. Bye.
-
