--- seq2seq-chatbot/data/twitter/pull
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-wget -c 'https://www.dropbox.com/s/tmfwptbs3q180p0/seq2seq.twitter.tar.gz?dl=0' -O seq2seq.twitter.tar.gz
