--- seq2seq-chatbot/data/twitter/pull_raw_data
@@ -1,2 +1,0 @@
-#!/bin/bash
-wget -c https://raw.githubusercontent.com/Marsan-Ma/chat_corpus/master/twitter_en.txt.gz
