--- seq2seq-chatbot/requirements.txt
@@ -1,7 +1,0 @@
-scikit-learn
-tensorflow
-tensorlayer
-numpy
-click
-tqdm
-nltk
